+++ proj/libs/uart/src/uart.c
@@ -212,7 +212,7 @@ int uart_set_parity(int base_addr, uart_parity par){
     conf = (conf & (~UART_PARITY)) | parity;
     return uart_set_lcr(base_addr, conf);
 }
-int uart_set_bit_rate(int base_addr, float bit_rate){
+int uart_set_bit_rate(int base_addr, double bit_rate){
     int ret = SUCCESS;
     uint16_t latch = UART_BITRATE/bit_rate;
     uint8_t dll = UART_GET_DLL(latch);
